requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
dataclasses==0.6
Django==4.2.7
django-cors-headers==4.3.0
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
idna==3.4
payme-pkg==2.6.3
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
